requirements.txt
mcp
httpx
uvicorn
starlette
python-dotenv

